+++ ibis/backends/pandas/aggcontext.py
