+++ ibis/backends/pandas/execution/constants.py
