+++ ibis/backends/pandas/execution/selection.py
