+++ ibis/backends/pandas/execution/strings.py
