+++ ibis/backends/pandas/execution/structs.py
