+++ ibis/backends/pandas/execution/timecontext.py
