+++ ibis/backends/pandas/execution/util.py
