+++ ibis/backends/pandas/execution/window.py
@@ -12,10 +12,10 @@ from pandas.core.groupby import SeriesGroupBy
 import ibis.common.exceptions as com
 import ibis.expr.operations as ops
 import ibis.expr.window as win
-import ibis.pandas.aggcontext as agg_ctx
 from ibis.expr.scope import Scope
 from ibis.expr.timecontext import TIME_COL
 from ibis.expr.typing import TimeContext
+from .. import aggcontext as agg_ctx
 from ..aggcontext import AggregationContext
 from ..core import (
     compute_time_context,
