+++ ibis/backends/pandas/tests/execution/test_operations.py
