+++ ibis/backends/pandas/tests/execution/test_strings.py
