+++ ibis/backends/pandas/tests/execution/test_structs.py
