+++ ibis/backends/pandas/tests/test_aggcontext.py
