+++ ibis/backends/pandas/tests/test_client.py
