+++ ibis/backends/pandas/tests/test_core.py
