+++ ibis/backends/pandas/tests/test_dispatcher.py
