+++ ibis/backends/pandas/trace.py
