+++ ibis/file/hdf5.py
@@ -3,7 +3,7 @@ import pandas as pd
 import ibis.expr.operations as ops
 import ibis.expr.schema as sch
 from ibis.file.client import FileClient
-from ibis.pandas.core import execute, execute_node
+from ibis.backends.pandas import execute, execute_node
 
 
 def connect(path):
