+++ ibis/file/parquet.py
@@ -10,8 +10,8 @@ import ibis.expr.operations as ops
 import ibis.expr.schema as sch
 import ibis.expr.types as ir
 from ibis.file.client import FileClient
-from ibis.pandas import PandasDialect
-from ibis.pandas.core import execute, execute_node
+from ibis.backends.pandas import PandasDialect
+from ibis.backends.pandas import execute, execute_node
 
 dialect = PandasDialect
