+++ ibis/tests/all/test_temporal.py
@@ -8,7 +8,8 @@ from pytest import param
 
 import ibis
 import ibis.expr.datatypes as dt
-from ibis.pandas.execution.temporal import day_name
+from ibis.backends.pandas import execute
+from ibis.backends.pandas.execution.temporal import day_name
 from ibis.tests.backends import (
     BigQuery,
     Clickhouse,
@@ -357,7 +358,7 @@ def test_to_timestamp(backend, con, unit):
     factor = unit_factors[unit]
 
     ts = ibis.timestamp('2018-04-13 09:54:11.872832')
-    pandas_ts = ibis.pandas.execute(ts).floor(unit).value
+    pandas_ts = execute(ts).floor(unit).value
 
     # convert the now timestamp to the input unit being tested
     int_expr = ibis.literal(pandas_ts // factor)
